+++ VGP337/KMeans/GameState.cpp
@@ -40,7 +40,7 @@ namespace
 
 		CR.assignment.clear();
 		CR.assignment.resize(data.size());
-		for (int i = 0; i < maxIteration; ++i)
+		for (size_t i = 0; i < maxIteration; ++i)
 		{
 			std::vector<Vector3> newCentroids;
 			std::vector<size_t> clusterEntryCount;
